Add support for Alpine edge
