+++ sources/alpine-http.go
@@ -23,11 +23,33 @@ func NewAlpineLinuxHTTP() *AlpineLinuxHTTP {
 
 // Run downloads an Alpine Linux mini root filesystem.
 func (s *AlpineLinuxHTTP) Run(definition shared.Definition, rootfsDir string) error {
-	fname := fmt.Sprintf("alpine-minirootfs-%s-%s.tar.gz", definition.Image.Release,
+	releaseFull := definition.Image.Release
+	releaseShort := releaseFull
+
+	if definition.Image.Release == "edge" {
+		if definition.Source.SameAs == "" {
+			return fmt.Errorf("You can't use Alpine edge without setting same_as")
+		}
+
+		releaseFull = definition.Source.SameAs
+		releaseShort = releaseFull
+	}
+
+	releaseField := strings.Split(releaseFull, ".")
+	if len(releaseField) == 2 {
+		releaseShort = fmt.Sprintf("v%s", releaseFull)
+		releaseFull = fmt.Sprintf("%s.0", releaseFull)
+	} else if len(releaseField) == 3 {
+		releaseShort = fmt.Sprintf("v%s.%s", releaseField[0], releaseField[1])
+	} else {
+		return fmt.Errorf("Bad Alpine release: %s", releaseFull)
+	}
+
+	fname := fmt.Sprintf("alpine-minirootfs-%s-%s.tar.gz", releaseFull,
 		definition.Image.ArchitectureMapped)
-	tarball := fmt.Sprintf("%s/v%s/releases/%s/%s", definition.Source.URL,
-		strings.Join(strings.Split(definition.Image.Release, ".")[0:2], "."),
-		definition.Image.ArchitectureMapped, fname)
+
+	tarball := fmt.Sprintf("%s/%s/releases/%s/%s", definition.Source.URL,
+		releaseShort, definition.Image.ArchitectureMapped, fname)
 
 	url, err := url.Parse(tarball)
 	if err != nil {
@@ -63,6 +85,26 @@ func (s *AlpineLinuxHTTP) Run(definition shared.Definition, rootfsDir string) er
 	err = lxd.Unpack(filepath.Join(os.TempDir(), fname), rootfsDir, false, false)
 	if err != nil {
 		return err
+	}
+
+	if definition.Image.Release == "edge" {
+		// Upgrade to edge
+		exitChroot, err := shared.SetupChroot(rootfsDir)
+		if err != nil {
+			return err
+		}
+
+		err = shared.RunCommand("sed", "-i", "-e", "s/v[[:digit:]]\\.[[:digit:]]/edge/g", "/etc/apk/repositories")
+		if err != nil {
+			return err
+		}
+
+		err = shared.RunCommand("apk", "upgrade", "--update-cache", "--available")
+		if err != nil {
+			return err
+		}
+
+		exitChroot()
 	}
 
 	return nil
